Add files via upload
This branch contains Assignment 3-2/5/24 

which contains these questions

Search an Element in a Linked List:
Explain the iterative approach to search for an element in a linked list. Provide a step-by-step explanation.
Compare and contrast the iterative and recursive approaches for searching an element in a linked list. What are the advantages and disadvantages of each approach? Provide a code and explanation
 2. Find Length of a Linked List:

Describe how you would find the length of a linked list using an iterative approach. Provide a code and explanation.
Discuss the recursive approach to finding the length of a linked list. How does it work, and what are the termination conditions? Provide a code and explanation.
3. Reverse a Linked List:

Implement an iterative function to reverse a linked list. Provide the code and explain how it works.
Explore the recursive method for reversing a linked list. How does it differ from the iterative approach, and what are the potential challenges? Provide a code and explanation.